Implement "FetchStyle" for ascii post processing
ASCII art is defined as a palette of four `#$%&` characters to which
a particular "style" escape code is applied depending on the
palette index.

Allows for things like "width" to properly be calculated from the
ascii art while still being able to apply colors
and possibly a non-color style to the art, or leaving it as plain ascii.